#1 Savepoint

+++ g2config/g2config.h
@@ -1,9 +1,10 @@
-#include <stdlib.h>
+#include <stdint.h>
 #include <stdio.h>
+#include <stdlib.h>
 #include "libg2config.h"
 
 typedef void* ExportHandle;
 typedef void*(*resize_buffer_type)(void *, size_t);
 
 
-long long G2config_create_helper();
+uintptr_t G2config_create_helper();
